requirements.txt
Pillow==10.1.0
matplotlib==3.8.0